* Added percentages to progress for larger transferred files, improved speed and quality of the normals smoothing function, added more MP refreshes during entity object loading and slowed down transfer rate to be kinder to Photon

+++ GameGuru Core/Dark Basic Public Shared/Dark Basic Pro SDK/DarkSDKMore/PhotonMultiplayer/PhotonMultiplayer.cpp
@@ -688,7 +688,7 @@ int PhotonAmIFileSynced()
 	return 0;
 }
 
-int PhotonGetFileProgress()
+float PhotonGetFileProgress()
 {
 	if ( g_pPhotonView )
 	{
